[cdc] Refactor codes cdcTimestampExtractor to separate different cdc dataSources.
